requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.10
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
streamlit>=1.30.0
pandas>=2.0.0
pydantic>=2.5.0
